\name{PBSsatellite}
\alias{PBSsatellite-package}
\title{Plotting and Statistical Analysis of Satellite Data}
\description{
  This software creates a standardize object of NetCDF data know as
  \code{\link{ncdfData}}. The \code{\link{ncdfData}} object allows for multiple
  NetCDF files to become compatible, regardless of differences in
  length, attributes, data type, and version.
  
  The \code{\link{ncdfData}} object allows users to manipulate satellite data
  for their own usage such as scaling, region clipping, and removing erroneous
  data values. PBSsatellite also provides easy access to satellite data such
  as extracting data of interested based on dates and date ranges. Users have
  the ability to run summary functions on \code{\link{ncdfData}} which allows
  for statistical analysis of specific regions of data over time. Users also
  have the ability to create their own summary functions for more specific
  staistical analysis.}
\seealso{
  \code{\link{PBSmapping}}
}
\author{
  \href{https://ca.linkedin.com/in/nicholas-lefebvre-b94b21132}{Nicholas Lefebvre}, Software Developer, Edmonton AB\cr
  \href{https://academic.macewan.ca/boersn/}{Nicholas M. Boers}, Associate Professor -- Computer Science\cr
  MacEwan University, Edmonton AB\cr
  Last modified \code{Rd: 2015-04-13}
}
\docType{package}
