\name{sst}
\docType{data}
\alias{sst}
\title{Long Term Mean of Sea Surface Temperature}
\description{
  This is a \code{\link{ncdfData}} object used in the PBSsatellite
  examples. It contains a data set of long term means of sea surface
  temperature for several months of 2001. This data set has grid spacing
  of 1.0 degree latitude and 1.0 degree longitude.
}
\note{
  Names attribute on this ncdfData object were renamed
  from 01-MM-DD to 2001-MM-DD for easier understanding
  in example code.
}
\format{\code{\link{ncdfData}}}
\source{NOAA_OI_SST_V2 data provided by the NOAA/OAR/ESRL PSD, Boulder,
  Colorado, USA, from their Web site at \url{http://www.esrl.noaa.gov/psd/}.
}
\references{Reynolds, R.W., N.A. Rayner, T.M. Smith, D.C. Stokes, and
  W. Wang (2002) An improved in situ and satellite SST analysis for
  climate. \emph{J. Climate} \bold{15}: 1609--1625.
}
\author{
  \href{https://ca.linkedin.com/in/nicholas-lefebvre-b94b21132}{Nicholas Lefebvre}, Software Developer, Edmonton AB\cr
  Last modified \code{Rd: 2015-04-10}
}

\keyword{datasets}
